requirements.txt
torch
transformers
sentence-transformers
huggingface-hub
faiss-cpu
numpy
scikit-learn
tqdm
wikipedia-api